Rename files to pascal case
