+++ src/components/PersonCard/PersonCard.module.scss
