+++ src/sections/Team/Team.module.scss
